refactor(begemotic): Rename begemotic directory, move enums and GeoCharacteristicsSchema to a separate files
